Remove Timekeeper.Toolbox dependency
Spurred by a strong desire to vastly improve overall TK performance, I'm
bringing the DBI code into the project itself. Not that the DLL is the
issue (it's not) but this will make it easier to work with, debug, test,
and deploy.

Brought over four core modules: DBI, Common, Log, and UUID.

This will also help position me for future development, whether this
thing ends up on the web or gets ported to macOS. A nice, self-contained
system will be easier to port over dealing with assemblies.

Plus, I have an extreme dislike now for Technitivity.Toolbox. It needs
to be completely decoupled from itself. It's architected terribly.