Commit final changes to 2.0.701.1
This is from May 29, 2017. I'm not sure why I didn't commit this
at the time. I just got Timekeeper working on macOS this week, using
CodeWeavers CrossOver and it's been mostly successful. The next
version of TK development will focus on key bug fixes as well as
optimizations for this "new" platform.