Color older entries when timer is running [Mantis #1442]
This helps ensure the user isn't editing an old entry when they
think they're on the current entry. I've done this many times: gone
back to something old and then forgot about it. Later, I come back
to close out the "current" entry but I'm on the old one.

The ticket said to give the indicator to the current entry, but
after some testing, it's more intuitive to do it this way.