Change "New" keyboard shortcuts
Before: Ctrl+N was used to create a new Timekeeper database and
Ctrl+Shift+N was used to create a new log entry in the journal. That
is not a good OOTB experience.

Now: Ctrl+N creates a new entry. There is no shortcut for creating
a new Timekeeper database. It's a rare event and does not need a
shortcut. (The user can assign a custom shortcut if necessary.)

Mantis #1369