M#1405: Commit initial WIP for a "join entry" function
The actual task was: add a delete entry function. But the primary
use case for delete functionality is over-splitting an entry. So,
I do get a delete out of it, but it also serves as unsplit.

This commit contains all the wiring for it, but does not actually
perform the join.

I'm also not going to commit anything without a ticket number any
longer. This commit represents my first use of the M#9999 notation,
meaning "Mantis Ticket Number 9999"