Commit 3.2 work
Okay, this is a bit of a mess. And, yes, it's my first commit in
more than four years. Back in 2021, it looks like I started working
on a 3.2 version and in 2024, I "released" 3.2.103.1. But I never
committed any of the in-between work.

As I put Timekeeper 5 plans on hold (and continue to use ATracker
as my mobile data-collection app), I'd like to create a TK 3.3
line to help me bridge this gap. See Jira TK-478 for details on the
plan.

For now, here's what's changed since March 2021:

* Began working on the idea of a "current location"
* Updated the release date to "Spring 2024"
* Put a band-aid on the worst TK3 bug ever: exception calculating seconds
* Some additional debugging, etc.

Maybe more, but that's the gist of it. Just committing this as "a
point in time" so that I can continue forward with the ATracker+TK5
project in the future.